+++ apps/target-service/src/target-service.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class AppService {
+export class TargetServiceService {
   getHello(): string {
     return 'Hello World!';
   }
