+++ nest-cli.json
@@ -1,8 +1,32 @@
 {
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
-  "sourceRoot": "src",
+  "sourceRoot": "apps/snapmatch/src",
   "compilerOptions": {
-    "deleteOutDir": true
+    "deleteOutDir": true,
+    "webpack": true,
+    "tsConfigPath": "apps/snapmatch/tsconfig.app.json"
+  },
+  "monorepo": true,
+  "root": "apps/snapmatch",
+  "projects": {
+    "snapmatch": {
+      "type": "application",
+      "root": "apps/snapmatch",
+      "entryFile": "main",
+      "sourceRoot": "apps/snapmatch/src",
+      "compilerOptions": {
+        "tsConfigPath": "apps/snapmatch/tsconfig.app.json"
+      }
+    },
+    "target-service": {
+      "type": "application",
+      "root": "apps/target-service",
+      "entryFile": "main",
+      "sourceRoot": "apps/target-service/src",
+      "compilerOptions": {
+        "tsConfigPath": "apps/target-service/tsconfig.app.json"
+      }
+    }
   }
 }
